Delete sandcat.go-windows
Remove this file because the payload already exists in the sandcat plugin, and having two payloads causes a conflict when trying to dynamically compile. Essentially, the dynamically compiled payload gets stored in the sandcat plugin's payload directory, but sometimes the caldera server will pull down the static version stored in this directory since it looks for the first file path.