Rename "Build" workflow to "Build Micro Plans"
This workflow originated in a standalone micro plans project, but now
that it's part of this larger project, the new name clarifies what this
workflow is for.

Also restrict the workflow to only build micro artifacts if some file in
the micro_emulation_plans/ tree was modified -- this reduces wasted
build time and storage when updating macro plans.